requirements.txt
fastapi==0.79.0
jose==1.0.0
passlib==1.7.4
pydantic==1.9.0
python-dotenv==1.0.0
python_jose==3.3.0
SQLAlchemy==1.4.39
alembic==1.11.3
pytest==7.4.0
uvicorn==0.23.2
psycopg2-binary
